ingestor/requirements.txt
paho-mqtt==2.1.0
redis==5.0.8
